pyproject.toml
[project]
name = "code_review_agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Mykhailo",email = "mykhailo.ulanovych@exilion.com"}
]
readme = "README.md"
packages = [
    { include = "code_review_agent", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.12.3"}
openai = "^1.23.6"
instructor = "^1.2.2"
gitpython = "^3.1.43"
pyyaml = "^6.0.1"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
code-review-agent = "code_review_agent.cli:app"